+++ com.aobuchow.sample.commander/fragment.e4xmi
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="ASCII"?>
 <fragment:ModelFragments xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:basic="http://www.eclipse.org/ui/2010/UIModel/application/descriptor/basic" xmlns:fragment="http://www.eclipse.org/ui/2010/UIModel/fragment" xmi:id="_BxaXACerEeWxCPrV0pAZQQ">
   <fragments xsi:type="fragment:StringModelFragment" xmi:id="_EI7oACerEeWxCPrV0pAZQQ" featurename="descriptors" parentElementId="xpath:/">
-    <elements xsi:type="basic:PartDescriptor" xmi:id="_KMklcCerEeWxCPrV0pAZQQ" elementId="test.partDescFragment.ASampleE4View" label="File Manager" iconURI="platform:/plugin/com.aobuchow.sample.commander/icons/Sample.png" tooltip="" category="Sample Commander" contributionURI="bundleclass://com.aobuchow.sample.commander/com.aobuchow.sample.commander.parts.FileManagerEditor">
+    <elements xsi:type="basic:PartDescriptor" xmi:id="_KMklcCerEeWxCPrV0pAZQQ" elementId="test.partDescFragment.ASampleE4View" label="File Manager" iconURI="platform:/plugin/com.aobuchow.sample.commander/icons/Sample.png" tooltip="" category="Sample Commander" contributionURI="bundleclass://com.aobuchow.sample.commander/com.aobuchow.sample.commander.editor.FileManagerEditor">
       <tags>View</tags>
       <tags>categoryTag:Sample Commander</tags>
     </elements>
