+++ com.aobuchow.sample.commander/src/com/aobuchow/sample/commander/editor/ContainerEditorInput.java
@@ -1,4 +1,4 @@
-package com.aobuchow.sample.commander.parts;
+package com.aobuchow.sample.commander.editor;
 
 import java.net.URI;
