+++ com.aobuchow.sample.commander/src/com/aobuchow/sample/commander/editor/ContainerLinkHelper.java
@@ -1,4 +1,4 @@
-package com.aobuchow.sample.commander;
+package com.aobuchow.sample.commander.editor;
 
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.viewers.StructuredSelection;
@@ -6,9 +6,6 @@ import org.eclipse.ui.IEditorInput;
 import org.eclipse.ui.IEditorPart;
 import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.navigator.ILinkHelper;
-
-import com.aobuchow.sample.commander.parts.FileManagerEditor;
-import com.aobuchow.sample.commander.parts.IContainerEditorInput;
 
 public class ContainerLinkHelper implements ILinkHelper {
