+++ com.aobuchow.sample.commander/src/com/aobuchow/sample/commander/handlers/OpenResourceHandler.java
@@ -1,4 +1,4 @@
-package com.aobuchow.sample.commander.parts;
+package com.aobuchow.sample.commander.handlers;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
@@ -21,6 +21,8 @@ import org.eclipse.ui.PartInitException;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.handlers.HandlerUtil;
 import org.eclipse.ui.ide.IDE;
+
+import com.aobuchow.sample.commander.editor.ContainerEditorInput;
 
 public class OpenResourceHandler extends AbstractHandler {
